requirements.txt
requests
pillow
flask[all]
idna
unicorn